chore(SDK-4215): Moves out inner classes
