+++ clevertap-core/src/main/java/com/clevertap/android/sdk/cryption/MigrationResult.kt
@@ -1,0 +1,6 @@
+package com.clevertap.android.sdk.cryption
+
+data class MigrationResult(
+    val data: String,
+    val migrationSuccessful: Boolean
+)
